Add files via upload
This project is one where I decided to analyze something I'm passionate about. Videogames! I wanted to find out about not only the most sold game of all, but which companies sold the most and which consoles were the most profitable from the late 80's all the way up to the early 2010's